+++ gemelli/q2/_transformer.py
@@ -1,17 +1,8 @@
-import qiime2
-import pandas as pd
 import numpy as np
+import pandas as pd
+from qiime2 import Metadata
 from .plugin_setup import plugin
 from ._format import TrajectoryFormat
-
-
-def _read_dataframe(fh):
-    # Using `dtype=object` and `set_index` to avoid type casting/inference
-    # of any columns or the index.
-    df = pd.read_csv(fh, sep='\t', header=0)
-    df.set_index(df.columns[0], drop=True, append=False, inplace=True)
-    df.index = df.index.astype(str)
-    return df
 
 
 @plugin.register_transformer
@@ -24,11 +15,11 @@ def _1(data: pd.DataFrame) -> (TrajectoryFormat):
 
 @plugin.register_transformer
 def _2(ff: TrajectoryFormat) -> (pd.DataFrame):
-    with ff.open() as fh:
-        return _read_dataframe(fh)
+    # with ff.open() as fh:
+    return Metadata.load(str(ff)).to_dataframe()
 
 
 @plugin.register_transformer
-def _3(ff: TrajectoryFormat) -> (qiime2.Metadata):
-    with ff.open() as fh:
-        return qiime2.Metadata(_read_dataframe(fh))
+def _3(ff: TrajectoryFormat) -> (Metadata):
+    # with ff.open() as fh:
+    return Metadata.load(str(ff))