+++ gemelli/tests/test_factorization.py
@@ -72,9 +72,18 @@ class TestTenAls(unittest.TestCase):
         self.U4 = U4
         self.U5 = U5
 
+    def test_TenAls_center(self):
+        # TensorFactorization no noise w/ centered
+        TF = TensorFactorization(center=True).fit(self.TE)
+        L1, L2, L3 = TF.loadings
+        for l1v in L1.mean(axis=1):
+            self.assertLess(l1v, 1)
+        for l2v in L2.mean(axis=1):
+            self.assertLess(l2v, 1)
+
     def test_TenAls_noiseless(self):
         # TensorFactorization no noise
-        TF = TensorFactorization().fit(self.TE)
+        TF = TensorFactorization(center=False).fit(self.TE)
         L1, L2, L3 = TF.loadings
         s = TF.eigvals
         # test accuracy
@@ -94,18 +103,18 @@ class TestTenAls(unittest.TestCase):
 
     def test_TenAls_mode4_noiseless(self):
         # TODO check values
-        TF = TensorFactorization().fit(self.TE4)
+        TF = TensorFactorization(center=False).fit(self.TE4)
         L1, L2, L3, L4 = TF.loadings
         # test accuracy
 
     def test_TenAls_mode5_noiseless(self):
         # TODO check values
-        TF = TensorFactorization().fit(self.TE5)
+        TF = TensorFactorization(center=False).fit(self.TE5)
         L1, L2, L3, L4, L5 = TF.loadings
 
     def test_TenAls_noise(self):
         # TensorFactorization no noise
-        TF = TensorFactorization().fit(self.TE_noise)
+        TF = TensorFactorization(center=False).fit(self.TE_noise)
         L1, L2, L3 = TF.loadings
         s = TF.eigvals
         # test accuracy
@@ -125,12 +134,12 @@ class TestTenAls(unittest.TestCase):
 
     def test_TenAls_mode4_noise(self):
         # TODO check values
-        TF = TensorFactorization().fit(self.TE_noise4)
+        TF = TensorFactorization(center=False).fit(self.TE_noise4)
         L1, L2, L3, L4 = TF.loadings
 
     def test_TenAls_mode5_noise(self):
         # TODO check values
-        TF = TensorFactorization().fit(self.TE_noise5)
+        TF = TensorFactorization(center=False).fit(self.TE_noise5)
         L1, L2, L3, L4, L5 = TF.loadings
 
     def test_khatri_rao(self):
@@ -155,17 +164,17 @@ class TestTenAls(unittest.TestCase):
     def test_errors(self):
         # test not array
         with self.assertRaises(ValueError):
-            TensorFactorization().fit(list(range(10)))
+            TensorFactorization(center=False).fit(list(range(10)))
         # test if none missing
         with self.assertRaises(ValueError):
-            TensorFactorization().fit(np.ones((5, 5)))
+            TensorFactorization(center=False).fit(np.ones((5, 5)))
         # test no nan(s)
         TE_errors = self.TE
         TE_errors[0, :, :] = np.inf
         with self.assertRaises(ValueError):
-            TensorFactorization().fit(TE_errors)
+            TensorFactorization(center=False).fit(TE_errors)
         # test max rank
         with self.assertRaises(ValueError):
-            TensorFactorization(
-                n_components=np.max(
-                    self.TE_noise.shape) +10).fit(self.TE_noise)
+            paeram_ = np.max(self.TE_noise.shape) + 10
+            TensorFactorization(center=False,
+                                n_components=paeram_).fit(self.TE_noise)